graph_nn_practice/requirements.txt
torch-geometric
torch-sparse
torch-scatter
torch-cluster